permisos colab

+++ src/routes/tactos/cab/+page.svelte
@@ -15,11 +15,13 @@
     import {guardarHistorial} from "$lib/historial/lib"
     import RadioButton from "$lib/components/RadioButton.svelte";
     import { createPer } from "$lib/stores/permisos.svelte";
-    import { getPermisosList } from "$lib/permisosutil/lib";
+    
     import permisos from "$lib/stores/permisos";
     import { goto } from "$app/navigation";
     import MultiSelect from "$lib/components/MultiSelect.svelte";
     import { getEstadoNombre,getEstadoColor } from "$lib/components/estadosutils/lib";
+    //Permisos
+    import{getPermisosList,getPermisosMessage} from "$lib/permisosutil/lib"
     //Actualizacion
     import { actualizacion,deboActualizar } from '$lib/stores/offline/actualizar';
     import { customoffliner } from '$lib/stores/offline/custom.svelte';
@@ -35,7 +37,7 @@
     import { getComandosSQL, setComandosSQL, flushComandosSQL} from '$lib/stores/sqlite/dbcomandos';
     import {getTotalSQL,setTotalSQL,setUltimoTotalSQL} from "$lib/stores/sqlite/dbtotal"
     import {getUserOffline,setDefaultUserOffline} from "$lib/stores/capacitor/offlineuser"
-    import {getCabOffline,setDefaultCabOffline} from "$lib/stores/capacitor/offlinecab"
+    import {getCabOffline,setDefaultCabOffline,updatePermisos} from "$lib/stores/capacitor/offlinecab"
     import {
         updateLocalTactosSQLUser,
         
@@ -61,6 +63,7 @@
     let getlocal = $state(false)
     let getvelocidad = $state(0)
     let getactualizacion = $state(0)
+    let getpermisos = $state("")
     let cargado = $state(false)
     let caber = createCaber()
     let cab = caber.cab
@@ -181,6 +184,7 @@
         }
     }
     function eliminarOnline() {
+        
         Swal.fire({
             title: 'Eliminar tacto',
             text: '¿Seguro que deseas eliminar el tacto?',
@@ -261,6 +265,13 @@
         let isOnline = await getOnlyInternet()
         intermitenter.addIntermitente(isOnline)
         if(coninternet.connected){
+            caboff = await updatePermisos(pb,usuarioid)
+            getpermisos = caboff.permisos
+            let listapermisos = getPermisosList(caboff.permisos)
+            if(!listapermisos[4]){
+                Swal.fire("Error permisos",getPermisosMessage(4),"error")
+                return 
+            }
             eliminarOnline()
         }
         else{
@@ -336,6 +347,10 @@
         
         onChangeTactos()
         filterUpdate()
+
+        caboff = await updatePermisos(pb,usuarioid)
+        getpermisos = caboff.permisos
+        
         cargado = true
     }
     async function getLocalSQL() {
@@ -478,6 +493,13 @@
         }
     }
     async function editarOnline() {
+        caboff = await updatePermisos(pb,usuarioid)
+        getpermisos = caboff.permisos
+        let listapermisos = getPermisosList(caboff.permisos)
+        if(!listapermisos[4]){
+            Swal.fire("Error permisos",getPermisosMessage(4),"error")
+            return 
+        }
         try{
             let data = {
                fecha:fecha +" 03:00:00" ,
